fix: restructure project to add pkg/email package
