--- letter.txt.go
@@ -1,1 +1,0 @@
-package main
