+++ main.go
@@ -2,18 +2,18 @@ package main
 
 import (
 	"log"
-	"pace-sender/utils"
+	"pace-sender/pkg/email"
 )
 
 func main() {
 	// Get the list of valid email addresses
-	validEmails, err := utils.GetValidEmails()
+	validEmails, err := email.GetValidEmails()
 	if err != nil {
 		log.Fatalf("Error validating emails: %v", err)
 	}
 
 	// Send emails only to valid email addresses
-	err = utils.SendEmailsToValidRecipients(validEmails)
+	err = email.SendEmailsToValidRecipients(validEmails)
 	if err != nil {
 		log.Fatalf("Error sending emails: %v", err)
 	}
